hydration function
The beginnings of a hydrate function courtesy of the Twit library.
The core can be found in app/utils/twitter which will read a segment
of tweet ids from a file and hdyrate them.

There is also the start a hydrate and updateProgress actions in
datasets that need to be scheduled to run on a hearbeat of somekind.
Also yet to be done is handling the actions that get generated and
updating the store appropriately to reflect that more data has been
hydrated.