hearbeat
Have heartbeat at regular intervals, but only kick off hydration when
there is not a rate limit in effect. This change was to help observed
behavior when the computer is put to sleep and then wakes up again, but
the time it was sleeping has passed. Regular heartbeats will mean that
Hydrator can re-evaluate whether fetching is possible on wakeup.