CSV
A dataset can be exported as CSV using a button that appears once a dataset
has been fully hydrated. The columns are hard coded at the moment and pulled
from a Tweet JSON -> CSV utility that has been built into Twarc for some time.
Ideas for improving the CSV are welcome!

Fixes #5